Refactor Python backend tests (#2945)
